tools/google-ads/requirements.txt
google-ads>=23.0.0
google-auth-oauthlib>=1.0.0
